refactor: extract common code and add auth to otel arrow api
